requirements.txt
# Core Trading Libraries
alpaca-trade-api==3.1.1
yfinance==0.2.28
ccxt==4.1.52  # Optional: For crypto trading

# Data Analysis & Machine Learning
pandas==2.0.3
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2
ta==0.10.2  # Basic technical analysis
# TA-Lib==0.4.28  # Advanced TA - requires C library installation

# Database & Storage
sqlalchemy==2.0.23
psycopg2-binary==2.9.9  # PostgreSQL support

# Web & API
requests==2.31.0
websocket-client==1.7.0
beautifulsoup4==4.12.3
lxml==5.1.0

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
schedule==1.2.0
pytz==2023.3
psutil==5.9.6

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.18.0

# Alerts & Notifications (choose one)
python-telegram-bot==20.7
# discord.py==2.3.2
# slack-sdk==3.26.1

# Development & Testing
pytest==7.4.3
pytest-cov==4.1.0
black==23.10.1
isort==5.13.2
flake8==6.1.0
pylint==3.0.3
autopep8==2.0.4
ipython==8.17.2